Switch to a label field, drop the input field
Reason: I think it's impossible or very hard to create and maintain an input
field that works for all situations.

Instead, I would rather have some wrapper for the label, and the input field
is added via slot. Like that, the control over the input is given back
to the component using it.